Conflicts:
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/GlobalTopicClient.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/GlobalTopicConstants.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/data/workorder/Schedule.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/data/workorder/ShipmentTracking.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/data/workorder/Workorder.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/data/workorder/WorkorderStatus.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/data/workorder/WorkorderSubstatus.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/rpc/server/HttpJsonBuilder.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/data/help/HelpTransactionHandler.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/data/photo/PhotoClient.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/data/workorder/ReportProblemType.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/data/workorder/WorkorderClient.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/data/workorder/WorkorderTransactionBuilder.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/data/workorder/WorkorderTransactionHandler.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/transaction/WebTransactionService.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/AuthActionBarActivity.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/AuthFragmentActivity.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/DrawerView.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/RateMeView.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/SignatureDisplayActivity.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/UnavailableCardView.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/ConfirmDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/CustomFieldDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/DurationDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/ExpiresDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/FeedbackDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/ReportProblemDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/ScheduleDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/dialog/ShipmentAddDialog.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/WorkorderActivity.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/WorkorderListFragment.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/detail/ActionBarTopView.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/detail/NotificationFragment.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/detail/NotificationView.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/detail/ShipmentListView.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/detail/TimeLogRowView.java
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/detail/WorkFragment.java
	projects/FieldNation/fieldnation/src/main/res/layout/dialog_add_shipment.xml
	projects/FieldNation/fieldnation/src/main/res/values/app_strings.xml
	projects/FieldNation/libUtils/src/main/java/com/fieldnation/utils/DateUtils.java
	projects/FieldNation/libUtils/src/main/java/com/fieldnation/utils/misc.java

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/data/workorder/ShipmentTracking.java
@@ -32,7 +32,7 @@ public class ShipmentTracking {
     @Json(name = "trackingId")
     private String _trackingId;
     @Json(name = "userId")
-    private Integer _userId;
+    private Long _userId;
     @Json(name = "workorderId")
     private Integer _workorderId;
     @Json(name = "workorderShipmentId")
@@ -85,7 +85,7 @@ public class ShipmentTracking {
         return _trackingId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return _userId;
     }
